+++ crates/sdk/src/program/mod.rs
@@ -4,6 +4,8 @@ pub mod arguments;
 pub mod core;
 /// Error types and definitions for program compilation, manipulation, and execution failures.
 pub mod error;
+/// Program execution's specific logger
+pub mod logger;
 /// Definitions and traits for resolving and satisfying execution witnesses for Simplicity programs.
 pub mod witness;
